Disk image has now been booted in FLEX and fixed up so that it looks more-or-less like a boot disk for a clean install. Specifically: 1/ deleted the STARTUP.SYS and STARTUP.TXT files that existed (these caused an error because they referenced another drive, which does not exist) 2/ deleted FLEX.COR (this is the base OS but was an older version than that present in the original FLEX.SYS) 3/ renamed the hacked FLEX.SYS as FLEX.COR
The following compromises have been made:
1/ per the 6809 FLEX Adaption Guide, FLEX.COR should start at t=1,s=1. On
this disk image, FLEX.COR starts at t=37,s=65. This means that QLOAD must
be modified (or patched) accordingly - the image location is in FCBs in the
first few bytes of the program.
2/ FLEX.COR need only be 18 sectors long but the image on this disk is 22
sectors long because of the way it was constructed. There are about 4 sectors
of unused space at the end of the image. Because of the FLEX binary file
format, this has no impact other that wasting a tiny bit of time (the adaption
guide (p22) even says the image is "over twenty sectors in length", which is
not quite accurate.