Change all processes to use rising edge clock. Previously they used a mixture of rising and falling, and this made the timing very tight (did not close timing with 50MHz clock). Moving everything to rising clock has no impact on functional performance; the pixels just come out 2 cycles later. Has a great beneficial effect on timing though. This effectively re-pipelines the output which required one counter comparator to change.
Fix functional bug brought to light by CUBIX. CUBIX does line endings as
LF CR rather than the usual CR LF. There was a bug in the display logic on
scroll: the line was cleared from the cursor position to the end rather than
from the start of line to the end. After a CR the cursor position was always
at the left so this bug was hidden. The effect of the bug was a wierd effect
on the bottom line which then scrolled up the screen. Took *ages* to work out
what was going on but, once spotted, was very straightforward to fix.