Screen 50:
This used to be a multicomp-specific screen but was just an exact copy of
screen 120, which was pointless and made maintenance more difficult. Remove
it.

Screen 45:

Reference screen 120 instead of screen 50; multicomp extras now start at
screen 50 instead of screen 51.

Screens 90, 91, 120:

Add new user variables BLK SCR (ANS Standard words) and BSTATE in preparation
for addition of BLOCK and BLOCK EXT words. Re-arrange the order of the user
variables so that those that require initialisation are first. Change the
initialisation table UINIT and its size #INIT to only initialise those words
that need it.

Change the default value of BLKADRS from 0x2.0A00 to 0x2.0800 (select the
first rather than the second block file by default).

Screen 109:

Extend definition of \ so that it accommodates BLK (if block file, skip
to next 64-byte boundary rather than to end of input region)

Screen 107:

Extend definition of EVALUATE so that it accommodates BLK (set BLK to 0).